+++ docs/changelog.rst
@@ -5,6 +5,7 @@ dev
 ---
 
 - **Breaking change:** :class:`~bridgekeeper.rules.Relation` and :class:`~bridgekeeper.rules.ManyRelation` no longer require the model class on the other side of the relation to be passed in as an argument.
+- **Breaking change:** :class:`~bridgekeeper.rules.ManyRelation` removes the ``attr`` argument, requiring only ``query_attr``.
 - **Breaking change:** Python 3.4 and Django 1.11 are no longer supported.
 
 0.7